+++ .idea/artifacts/autonomous_lawnmower.xml
@@ -1,0 +1,25 @@
+<component name="ArtifactManager">
+  <artifact type="javafx" name="autonomous_lawnmower">
+    <output-path>$PROJECT_DIR$/out/artifacts/autonomous_lawnmower</output-path>
+    <properties id="javafx-properties">
+      <options>
+        <option name="appClass" value="viewer.Main" />
+        <option name="description" value="" />
+        <option name="htmlPlaceholderId" value="" />
+        <option name="icons">
+          <JavaFxApplicationIcons>
+            <option name="windowsIcon" value="$PROJECT_DIR$/images/icon.ico" />
+          </JavaFxApplicationIcons>
+        </option>
+        <option name="title" value="IntelliMow" />
+        <option name="vendor" value="Daniel and Andriy" />
+        <option name="version" value="" />
+      </options>
+    </properties>
+    <root id="root">
+      <element id="archive" name="autonomous_lawnmower.jar">
+        <element id="module-output" name="autonomous-lawnmower" />
+      </element>
+    </root>
+  </artifact>
+</component>
